Move PlanNodeStats to upper level
